Prefix commands with bot's nickname
The ! operator is often used by other bots, and is kind of unfriendly,
anyway. I want zanegort to act at least a little bit more like a real
person. Thus, instead of prefixing commands with ! (or another chosen
prefix), commands will be prefixed with the bot's nickname. In other
words, they will take the form

    > zane, <cmd>

or

    > zane: <cmd>

Closes #7.