Plugin modules should adopt behaviour
Right now, plugins are hardcode into zane_bot. Eventually, I hope to
have a better plugin architecture; this is a first step towards that.

(I was considering creating a src/plugin directory and putting all
plugins in there, but I'm not sure that is kosher in the Erlang world;
other projects don't seem to create subdirectories under src/. Maybe the
plugin stuff should really be a separate app, but I can deal with that
issue later.)

Closes #8.