plug_ctcp: Don't respond if /msg'd instead of /ctcp'd
The logic here is a bit complicated, but if the stripped Head is not
lesser in length than Head, it means that no characters were stripped,
which means that the message (Head) was not delimited with \001, which
means it was _not_ a CTCP command and was (probably) a /msg instead. In
that case, don't bother responding.