"Rebarize" application
The application isn't _really_ rebarized; or rather, it's not _really_
an OTP-client application, and I'm not actually building a release or
anything like that. It's still started in the old-fashioned way (from
the shell). Combining some of the automatic flags from [1] with the
flags I can use to start an application described in [2], I _might_ be
able to get closer to producing a "real" application, but for now it's
still just more of an ad-hoc thing that can be started from the shell.
First steps, though!

  [1] http://www.howistart.org/posts/erlang/1
  [2] http://dozzie.jogger.pl/2014/02/07/erlang-running-your-code/