Compile zanegort using relx
relx is a tool that works in conjunction with Rebar to compile Erlang
releases.

Again, this isn't much of a "release". zanegort is not yet a standalone
application; executing the startup script, as per the README, simply
dumps you into a shell where you can call `zane_bot:start_link()` to
start up the bot. As a next step, I could possibly use the flags from
[1], in conjunction with the VM args as described in [2], to start the
app right away using the `_rel` script...but what I really should be
doing is eventually building a _real_ Erlang application. Still working
on that bit, though.

Anyway, at the very least, this should make it easier to include
third-party applications and libraries in zanegort. I might eventually
want to connect to a postgres database, for example, and this work could
help facilitate that.

  [1] http://dozzie.jogger.pl/2014/02/07/erlang-running-your-code/
  [2] http://www.howistart.org/posts/erlang/1