plug_cmd: Handle "http://" in URLs
Because of how the protocol lines are parsed, URLs with a protocol
component were broken up into a 2-element list. For example,
"http://www.example.com/" would be passed to the dispatch function as
["http", "//www.example.com"]. The dispatch function was not set up to
handle a URL as a list, so no match was made and the command failed.

With this commit, a protocol component in a URL will be properly
handled.

Closes #16.